Intellij auto-format the README
I keep doing it by habit muscle memory, and it always creates a big diff. This is noise so I revert it instead of piling it onto the existing work I'm doing. So now, when it's the only change in the commit, it can be done without distraction from other changes.